requirements.txt
opentelemetry-api==1.39.1
opentelemetry-instrumentation==0.42b0
opentelemetry-semantic-conventions==1.21.0